fix: schema check table exist before creation

+++ database/migrations/2024_09_01_101405_create_credit_scores_table.php
@@ -11,31 +11,33 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('credit_scores', function (Blueprint $table) {
-            $table->id();
+        if (! Schema::hasTable('credit_scores')) {
+            Schema::create('credit_scores', function (Blueprint $table) {
+                $table->id();
 
-            $table->string('identifier')->nullable()->unique(); // system generated e.g CSCORES-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e CSCORES-10MG-20240901-23
-            $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
-            $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
+                $table->string('identifier')->nullable()->unique(); // system generated e.g CSCORES-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e CSCORES-10MG-20240901-23
+                $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
+                $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
 
-            // linked the most active evaluation result used for the credit score check
-            $table->foreignId('txn_evaluation_id')->nullable()->constrained('credit_txn_history_evaluations')->onDelete('cascade');
+                // linked the most active evaluation result used for the credit score check
+                $table->foreignId('txn_evaluation_id')->nullable()->constrained('credit_txn_history_evaluations')->onDelete('cascade');
 
-            // since business rules can be set to active or not, the system need to pick the current business rules used to run this check
-            $table->json('business_rule_json')->nullable()->comment('store the current business rule used to perform this credit score that yield the credit_score_result');
-            $table->json('credit_score_result')->nullable()->comment('store result of the business rules comparison with txn history evaluation result');
+                // since business rules can be set to active or not, the system need to pick the current business rules used to run this check
+                $table->json('business_rule_json')->nullable()->comment('store the current business rule used to perform this credit score that yield the credit_score_result');
+                $table->json('credit_score_result')->nullable()->comment('store result of the business rules comparison with txn history evaluation result');
 
-            $table->float('score_percent')->nullable();
-            $table->float('score_value')->nullable();
-            $table->float('score_total')->nullable();
+                $table->float('score_percent')->nullable();
+                $table->float('score_value')->nullable();
+                $table->float('score_total')->nullable();
 
-            $table->foreignId('created_by_id')->nullable()->constrained('business_users')->onDelete('cascade');
+                $table->foreignId('created_by_id')->nullable()->constrained('business_users')->onDelete('cascade');
 
-            $table->enum('source', ['DASHBOARD', 'API']);
+                $table->enum('source', ['DASHBOARD', 'API']);
 
-            $table->unique(['business_id', 'customer_id', 'identifier'], 'score_business_customer_identifier_unique');
-            $table->timestamps();
-        });
+                $table->unique(['business_id', 'customer_id', 'identifier'], 'score_business_customer_identifier_unique');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
