+++ database/migrations/2024_09_01_113047_create_credit_repayment_debit_cards_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('credit_repayment_debit_cards', function (Blueprint $table) {
-            $table->id();
+        if (! Schema::hasTable('credit_repayment_debit_cards')) {
+            Schema::create('credit_repayment_debit_cards', function (Blueprint $table) {
+                $table->id();
 
-            $table->string('identifier')->nullable()->unique(); // system generated e.g DCRDS-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e DCRDS-10MG-20240901-23
-            $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
-            $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
+                $table->string('identifier')->nullable()->unique(); // system generated e.g DCRDS-VENDOR_CODE-YEARMONTHDAY-PRIMARY_ID i.e DCRDS-10MG-20240901-23
+                $table->foreignId('business_id')->constrained('businesses')->onDelete('cascade'); // BusinessType::VENDOR
+                $table->foreignId('customer_id')->constrained('credit_customers')->onDelete('cascade');
 
-            // ...mandate information
-            // TODO: Check paystack documentation to implement auto-debit card authorisation and add additional columns
+                // ...mandate information
+                // TODO: Check paystack documentation to implement auto-debit card authorisation and add additional columns
 
-            $table->index(['business_id', 'customer_id'], 'debit_business_customer_identifier_unique');
-            $table->timestamps();
-        });
+                $table->index(['business_id', 'customer_id'], 'debit_business_customer_identifier_unique');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
