+++ database/migrations/2024_09_01_114825_create_credit_repayment_schedules_table.php
@@ -11,20 +11,22 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('credit_repayment_schedules', function (Blueprint $table) {
-            $table->id();
+        if (! Schema::hasTable('credit_repayment_schedules')) {
+            Schema::create('credit_repayment_schedules', function (Blueprint $table) {
+                $table->id();
 
-            $table->foreignId('loan_id')->constrained('credit_loans')->onDelete('cascade');
-            $table->unsignedBigInteger('payment_id')->nullable()->index(); //repayment_log_id
+                $table->foreignId('loan_id')->constrained('credit_loans')->onDelete('cascade');
+                $table->unsignedBigInteger('payment_id')->nullable()->index(); //repayment_log_id
 
-            $table->decimal('amount', 18, 2)->nullable();
-            $table->decimal('late_fee', 18, 2)->nullable();
-            $table->datetime('due_date')->nullable();
+                $table->decimal('amount', 18, 2)->nullable();
+                $table->decimal('late_fee', 18, 2)->nullable();
+                $table->datetime('due_date')->nullable();
 
-            $table->string('payment_status')->nullable(); // PENDING, PAID
+                $table->string('payment_status')->nullable(); // PENDING, PAID
 
-            $table->timestamps();
-        });
+                $table->timestamps();
+            });
+        }
     }
 
     /**
